pyproject.toml
[project]
name = "fork-mfrc522-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pygame>=2.6.1",
    "rpi-gpio>=0.7.1",
    "setuptools>=80.9.0",
    "spidev>=3.7",
]
